refactor step memory and cleaning backward compactabilt
