--- src/traceml/diagnostics/step_memory_summary/context.py
@@ -1,15 +1,0 @@
-"""
-Compatibility shim for step-memory summary diagnostic contexts.
-"""
-
-from traceml.diagnostics.step_memory.adapters import (
-    StepMemorySummaryMetricSignals,
-    StepMemorySummaryTrendSignals,
-    build_step_memory_summary_signals,
-)
-
-__all__ = [
-    "StepMemorySummaryTrendSignals",
-    "StepMemorySummaryMetricSignals",
-    "build_step_memory_summary_signals",
-]
