--- src/traceml/diagnostics/step_memory_summary/rules.py
@@ -1,21 +1,0 @@
-"""
-Compatibility shim for step-memory summary diagnostic rules.
-"""
-
-from traceml.diagnostics.step_memory.rules import (
-    DEFAULT_STEP_MEMORY_SUMMARY_RULES,
-    CreepConfirmedRule,
-    CreepEarlyRule,
-    HighPressureRule,
-    ImbalanceRule,
-    run_step_memory_summary_rules,
-)
-
-__all__ = [
-    "HighPressureRule",
-    "ImbalanceRule",
-    "CreepConfirmedRule",
-    "CreepEarlyRule",
-    "DEFAULT_STEP_MEMORY_SUMMARY_RULES",
-    "run_step_memory_summary_rules",
-]
