--- src/traceml/reporting/sections/step_time/diagnosis.py
@@ -1,19 +1,0 @@
-"""Compatibility exports for step-time summary diagnosis."""
-
-from traceml.diagnostics.step_time.adapters import (
-    RankStepMetricSeries,
-    RankStepSignals,
-    build_summary_step_diagnosis,
-    build_summary_step_diagnosis_result,
-    diagnosis_result_to_json,
-    diagnosis_to_json,
-)
-
-__all__ = [
-    "RankStepMetricSeries",
-    "RankStepSignals",
-    "build_summary_step_diagnosis",
-    "build_summary_step_diagnosis_result",
-    "diagnosis_result_to_json",
-    "diagnosis_to_json",
-]
