--- src/traceml/reporting/summaries/step_time_diagnosis.py
@@ -1,13 +1,0 @@
-"""Compatibility exports for step-time summary diagnosis helpers."""
-
-from traceml.diagnostics.step_time.adapters import (
-    RankStepSignals,
-    build_summary_step_diagnosis,
-    build_summary_step_diagnosis_result,
-)
-
-__all__ = [
-    "RankStepSignals",
-    "build_summary_step_diagnosis",
-    "build_summary_step_diagnosis_result",
-]
